backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
alembic==1.13.1
pydantic[email]==1.10.14
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
strawberry-graphql==0.214.2
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
asgi-lifespan==2.1.0
